Address code review: portability, deadlocks, correctness, clarity
- apply_mask: replace compiler intrinsics with portable bit shifts
- WebSocketServer: fire on_disconnected_ outside mutex in all paths
  (send, broadcast, close_client, stop) to prevent callback deadlocks
- TerminalUI: handle SIGWINCH for terminal resize
- HTTP bind: fail fast on main thread instead of silently dying
- AOServer::dispatch: remove const_cast, take non-const reference
- next_player_number_/session_id: widen to uint64_t
- ProtocolRouter: log unexpected NX WebSocket data instead of silent drop
- AOPacketIDClient: rename fields to software/version, add dual-role comment
- NXServer::generate_token: use thread_local RNG to avoid data race
- NXServer::get_session_by_token: document O(n) for future indexing
- AOGameState::char_taken: change from int to bool (was truncating uint64_t)

+++ include/game/ServerSession.h
@@ -13,7 +13,7 @@
 
 struct ServerSession {
     uint64_t client_id = 0;    ///< WebSocket client ID (transport-level).
-    uint32_t session_id = 0;   ///< Unique server-assigned session ID.
+    uint64_t session_id = 0;   ///< Unique server-assigned session ID.
     std::string session_token; ///< Auth token (empty for AO legacy clients).
 
     std::string display_name; ///< Player display name / showname.
